+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Dev
 
+### Enhancements
+
+- Add `Iter.take_every/2` and `Iter.drop_every/2`
+
 ## v0.1.1 (2022-12-26)
 
 ### Enhancements